+++ src/engine.js
@@ -19,8 +19,11 @@
  */
 
 /* global window */
+var parser = require('../dist/parser');
+var shuffle = require('./shuffle');
+require('./date-extensions');
 
-var XPathJS = (function(){
+module.exports = (function(){
 	var XPathException,
 		XPathEvaluator,
 		XPathExpression,
@@ -2001,10 +2004,10 @@ var XPathJS = (function(){
 			
 			// Parse the expression
 			try {
-				tree = XPathJS._parser.parse(expression);
+				tree = parser.parse(expression);
 			} catch(err) {
 				message = 'The expression is not a legal expression.';
-				if (err instanceof XPathJS._parser.SyntaxError)
+				if (err instanceof parser.SyntaxError)
 				{
 					message += ' (line: ' + err.line + ', character: ' + err.column + ')';
 				}
@@ -5759,6 +5762,20 @@ var XPathJS = (function(){
 		XPathExpression: XPathExpression,
 		XPathResult: XPathResult,
 		XPathNamespace: XPathNamespace,
+		customXPathFunction: {
+			type: {
+				StringType: StringType,
+				NumberType: NumberType,
+				NodeSetType: NodeSetType,
+				BooleanType: BooleanType
+			},
+			add: function(name, fnObj){
+				functions[""][name] = fnObj;
+			},
+			remove: function(name) {
+				delete functions[""][name];
+			}
+		},
 		
 		/**
 		 * Get the current list of DOM Level 3 XPath window and document objects